requirements.txt
asn1crypto==1.4.0
azure-common==1.1.27
azure-core==1.17.0
azure-storage-blob==12.8.1
boto3==1.17.97
botocore==1.20.97
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
cryptography==3.4.8
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.21
numpy==1.20.3
oauthlib==3.1.1
oscrypto==1.2.1
pandas==1.2.4
pycparser==2.20
pycryptodomex==3.10.1
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
s3transfer==0.4.2
six==1.16.0
sniffio==1.2.0
snowflake-connector-python==2.6.2
urllib3==1.26.5
